requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
streamlit>=1.28.0
requests>=2.31.0
pyyaml>=6.0.1
